Use a map that stores the lookup keys for all entries per session. If there is no key for a given lookup then there is no need to make a network call to the Memcached server. Also the map is used to be able to clean all entries per http session when the session expires/is invalidated.

+++ src/main/java/org/wicketstuff/datastore/memcached/MemcachedDataStore.java
@@ -1,8 +1,10 @@
 package org.wicketstuff.datastore.memcached;
 
 import java.io.IOException;
-import java.io.Serializable;
 import java.net.InetSocketAddress;
+import java.util.Set;
+import java.util.SortedSet;
+import java.util.TreeSet;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.ConcurrentMap;
 import java.util.concurrent.TimeUnit;
@@ -20,11 +22,11 @@ import org.apache.wicket.util.time.Duration;
 public class MemcachedDataStore implements IDataStore
 {
 	/**
-	 * A prefix for the keys to avoid duplication of keys
+	 * A prefix for the keysPerSession to avoid duplication of keysPerSession
 	 * and to make it easier to find out who put the data
 	 * at the server
 	 */
-	private static final String KEY_PREFIX = "Wicket-Memcached";
+	private static final String KEY_SUFFIX = "Wicket-Memcached";
 
 	/**
 	 * A separator used for the key construction
@@ -42,13 +44,46 @@ public class MemcachedDataStore implements IDataStore
 	private final IMemcachedSettings settings;
 
 	/**
+	 * Tracks the keys for all operations per session.
+	 * Used to delete all entries for this session.
+	 */
+	private final ConcurrentMap<String, SortedSet<String>> keysPerSession =
+			new ConcurrentHashMap<String, SortedSet<String>>();
+	
+	/**
 	 * Constructor.
+	 *
+	 * Creates a MemcachedClient from the provided settings
 	 *
 	 * @param settings The configuration for the client
 	 */
 	public MemcachedDataStore(IMemcachedSettings settings)
 	{
+		this(createClient(settings), settings);
+	}
+
+	/**
+	 * Constructor.
+	 *
+	 * @param client   The connection to Memcached
+	 * @param settings The configuration for the client
+	 */
+	public MemcachedDataStore(MemcachedClient client, IMemcachedSettings settings)
+	{
+		this.client = Args.notNull(client, "client");
 		this.settings = Args.notNull(settings, "settings");
+	}
+
+	/**
+	 * Creates MemcachedClient with the provided hostname and port
+	 * in the settings
+	 *
+	 * @param settings  The configuration for the client
+	 * @return A MemcachedClient
+	 */
+	private static MemcachedClient createClient(IMemcachedSettings settings)
+	{
+		Args.notNull(settings, "settings");
 
 		String host = settings.getHost();
 		Checks.notEmptyShort(host, "host");
@@ -58,7 +93,7 @@ public class MemcachedDataStore implements IDataStore
 
 		try
 		{
-			this.client = new MemcachedClient(new InetSocketAddress(host, port));
+			return new MemcachedClient(new InetSocketAddress(host, port));
 		}
 		catch (IOException iox)
 		{
@@ -70,10 +105,11 @@ public class MemcachedDataStore implements IDataStore
 	public byte[] getData(String sessionId, int pageId)
 	{
 		byte[] bytes = null;
-		SessionData data = getSessionData(sessionId);
-		if (data != null)
+		String key = getKey(sessionId, pageId);
+		SortedSet<String> keys = keysPerSession.get(sessionId);
+		if (keys != null && keys.contains(key))
 		{
-			bytes = data.pages.get(pageId);
+			bytes = (byte[]) client.get(key);
 		}
 		return bytes;
 	}
@@ -81,34 +117,52 @@ public class MemcachedDataStore implements IDataStore
 	@Override
 	public void removeData(String sessionId, int pageId)
 	{
-		SessionData sessionData = getSessionData(sessionId);
-		if (sessionData != null)
+		Set<String> keys = keysPerSession.get(sessionId);
+		String key = getKey(sessionId, pageId);
+		if (keys != null && keys.contains(key))
 		{
-			byte[] removed = sessionData.pages.remove(pageId);
-			if (removed != null)
-			{
-				storeSessionData(sessionId, sessionData);
-			}
+			client.delete(key);
+			keys.remove(key);
 		}
 	}
 
 	@Override
 	public void removeData(String sessionId)
 	{
-		String key = getKey(sessionId);
-		client.delete(key);
+		Set<String> keys = keysPerSession.get(sessionId);
+		if (keys != null)
+		{
+			for (String key : keys)
+			{
+				client.delete(key);
+			}
+			keysPerSession.remove(sessionId);
+		}
 	}
 
 	@Override
 	public void storeData(String sessionId, int pageId, byte[] data)
 	{
-		SessionData sessionData = getSessionData(sessionId);
-		if (sessionData == null)
+		String key = getKey(sessionId, pageId);
+		SortedSet<String> keys = keysPerSession.get(sessionId);
+		if (keys == null)
 		{
-			sessionData = new SessionData();
+			keys = new TreeSet<String>();
+			SortedSet<String> old = keysPerSession.putIfAbsent(sessionId, keys);
+			if (old != null)
+			{
+				keys = old;
+			}
 		}
-		sessionData.pages.put(pageId, data);
-		storeSessionData(sessionId, sessionData);
+		keys.add(key);
+
+		Duration expirationTime = settings.getExpirationTime();
+
+		// TODO Improve to follow Memcached protocol.
+		// See net.spy.memcached.MemcachedClient.set(java.lang.String, int, java.lang.Object)()
+//		Time timeToExpire = Time.now().add(expirationTime);
+
+		client.set(key, (int) expirationTime.getMilliseconds(), data);
 	}
 
 	@Override
@@ -128,49 +182,27 @@ public class MemcachedDataStore implements IDataStore
 	public boolean canBeAsynchronous()
 	{
 		// no need to be asynchronous
+		// MemcachedClient is asynchronous itself
 		return false;
 	}
 
-
-	private String getKey(String sessionId)
-	{
-		StringBuilder key = new StringBuilder();
-		key
-			.append(KEY_PREFIX)
-			.append(SEPARATOR)
-			.append(sessionId);
-		return key.toString();
-	}
-
-	private SessionData getSessionData(String sessionId)
-	{
-		String key = getKey(sessionId);
-		return (SessionData) client.get(key);
-	}
-
 	/**
-	 * Stores the SessionData in Memcached.
+	 * Creates a key that is used for the lookup in Memcached.
+	 * The key starts with sessionId and the pageId so
+	 * {@linkplain #keysPerSession} can be sorted faster.
 	 *
-	 * @param sessionId  The session id
-	 * @param sessionData The pages' for this session
+	 * @param sessionId The id of the http session.
+	 * @param pageId    The id of the stored page
+	 * @return A key that is used for the lookup in Memcached
 	 */
-	private void storeSessionData(String sessionId, SessionData sessionData)
+	private String getKey(String sessionId, int pageId)
 	{
-		Duration expirationTime = settings.getExpirationTime();
-
-		// TODO Improve to follow Memcached protocol.
-		// See net.spy.memcached.MemcachedClient.set(java.lang.String, int, java.lang.Object)()
-//		Time timeToExpire = Time.now().add(expirationTime);
-
-		client.set(getKey(sessionId), (int) expirationTime.getMilliseconds(), sessionData);
-	}
-
-	/**
-	 * Holds the pages per session
-	 */
-	// TODO Get rid of this because it will keep not needed data until the session expires
-	private static class SessionData implements Serializable
-	{
-		ConcurrentMap<Integer, byte[]> pages = new ConcurrentHashMap<Integer, byte[]>();
+		return new StringBuilder()
+			.append(sessionId)
+			.append(SEPARATOR)
+			.append(pageId)
+			.append(SEPARATOR)
+			.append(KEY_SUFFIX)
+			.toString();
 	}
 }
